requirements.txt
pyqt6
pyqt6-webengine
ollama
requests
pyperclip
pyautogui
mss
pillow
markdown
openai
pypdf
pymupdf
